refactor: Segregates report controller from reporting supervision to add a new policy so that admins and instructors can access reports.
